nit: update .gitignore, udpate js-doc
