+++ backend/.gitignore
@@ -3,3 +3,4 @@ node_modules
 .env*
 *pnpm-lock.yaml
 *.log
+src/swagger_output.json
